Inline the demo data generator
